requirements-airflow.txt
# Dépendances de base
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.2

# Téléchargement de données financières
yfinance>=0.1.70

# MLOps
mlflow>=2.0.0
python-dotenv>=0.20.0

# Airflow providers
apache-airflow-providers-apache-kafka>=1.0.0
apache-airflow-providers-amazon>=6.0.0
apache-airflow-providers-http>=2.0.0
apache-airflow-providers-slack>=7.0.0

# Kafka
confluent-kafka>=1.9.2

# DVC
dvc>=2.10.0
dvc-s3>=2.10.0